requirements.txt
pytorch
torchvision
matplotlib
Pillow
hydra-core
wandb